+++ scripts/translate/mod.ts
@@ -58,5 +58,6 @@ await completeTranslationsFor("src/ui/svelte/user-menu/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/calendar-popup/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/media-popup/i18n/translations");
 await completeTranslationsFor("src/ui/svelte/notifications/i18n/translations");
+await completeTranslationsFor("src/ui/svelte/flyouts/i18n/translations");
 
 await completeTranslationsFor("src/background/i18n");